engine_traits: Add a macro for implementing RaftEngine for KvEngines
