+++ components/engine_panic/src/raft_engine.rs
@@ -2,118 +2,7 @@
 
 use crate::engine::PanicEngine;
 use crate::write_batch::PanicWriteBatch;
-use engine_traits::{Error, RaftEngine, RaftLogBatch, Result};
-use kvproto::raft_serverpb::RaftLocalState;
-use raft::eraftpb::Entry;
 
-impl RaftEngine for PanicEngine {
-    type LogBatch = PanicWriteBatch;
+use engine_traits::def_raft_engine;
 
-    fn log_batch(&self, capacity: usize) -> Self::LogBatch {
-        panic!()
-    }
-
-    fn sync(&self) -> Result<()> {
-        panic!()
-    }
-
-    fn get_raft_state(&self, raft_group_id: u64) -> Result<Option<RaftLocalState>> {
-        panic!()
-    }
-
-    fn get_entry(&self, raft_group_id: u64, index: u64) -> Result<Option<Entry>> {
-        panic!()
-    }
-
-    fn fetch_entries_to(
-        &self,
-        region_id: u64,
-        low: u64,
-        high: u64,
-        max_size: Option<usize>,
-        buf: &mut Vec<Entry>,
-    ) -> Result<usize> {
-        panic!()
-    }
-
-    fn consume(&self, batch: &mut Self::LogBatch, sync_log: bool) -> Result<usize> {
-        panic!()
-    }
-
-    fn consume_and_shrink(
-        &self,
-        batch: &mut Self::LogBatch,
-        sync_log: bool,
-        max_capacity: usize,
-        shrink_to: usize,
-    ) -> Result<usize> {
-        panic!()
-    }
-
-    fn clean(
-        &self,
-        raft_group_id: u64,
-        state: &RaftLocalState,
-        batch: &mut Self::LogBatch,
-    ) -> Result<()> {
-        panic!()
-    }
-
-    fn append(&self, raft_group_id: u64, entries: Vec<Entry>) -> Result<usize> {
-        panic!()
-    }
-
-    fn append_slice(&self, raft_group_id: u64, entries: &[Entry]) -> Result<usize> {
-        panic!()
-    }
-
-    fn put_raft_state(&self, raft_group_id: u64, state: &RaftLocalState) -> Result<()> {
-        panic!()
-    }
-
-    fn gc(&self, raft_group_id: u64, mut from: u64, to: u64) -> Result<usize> {
-        panic!()
-    }
-
-    fn purge_expired_files(&self) -> Result<Vec<u64>> {
-        panic!()
-    }
-
-    fn has_builtin_entry_cache(&self) -> bool {
-        panic!()
-    }
-
-    fn flush_metrics(&self, instance: &str) {
-        panic!()
-    }
-
-    fn reset_statistics(&self) {
-        panic!()
-    }
-
-    fn dump_stats(&self) -> Result<String> {
-        panic!()
-    }
-}
-
-impl RaftLogBatch for PanicWriteBatch {
-    fn append(&mut self, raft_group_id: u64, entries: Vec<Entry>) -> Result<()> {
-        panic!()
-    }
-
-    fn append_slice(&mut self, raft_group_id: u64, entries: &[Entry]) -> Result<()> {
-        panic!()
-    }
-
-    fn cut_logs(&mut self, raft_group_id: u64, from: u64, to: u64) {
-        panic!()
-    }
-
-    fn put_raft_state(&mut self, raft_group_id: u64, state: &RaftLocalState) -> Result<()> {
-        panic!()
-    }
-
-    fn is_empty(&self) -> bool {
-        panic!()
-    }
-}
+def_raft_engine!(PanicEngine, PanicWriteBatch);
